Namespace local proxy vars and refactor as factory
